+++ src/include/kyber_scalar/verify.h
@@ -7,10 +7,10 @@
 #include <stdint.h>
 
 #define verify KYBER_NAMESPACE(verify)
-int verify(const u8 *a, const u8 *b, size_t len);
+int verify(const u8 *a, const u8 *b, u64 len);
 
 #define cmov KYBER_NAMESPACE(cmov)
-void cmov(u8 *r, const u8 *x, size_t len, u8 b);
+void cmov(u8 *r, const u8 *x, u64 len, u8 b);
 
 #define cmov_int16 KYBER_NAMESPACE(cmov_int16)
 void cmov_int16(i16 *r, i16 v, u16 b);
